+++ src/Polynomial.cc
@@ -54,7 +54,7 @@ Polynomial Polynomial::operator*(const Polynomial& other){
 	return Polynomial(product, field.field_size) % this->field;
 }
 
-Polynomial Polynomial::pow(int power){
+Polynomial Polynomial::pow(long long power){
     return pow_mod(*this, power, this->field, Polynomial(0b1, field.field_size));
 }
 
@@ -62,10 +62,11 @@ Polynomial Polynomial::inverse(){
     long long order = field.field_size - 1;
     return pow(order-1);
 }
-Polynomial Polynomial::pow_mod(Polynomial base, int exp, IrreduciblePolynomial p, Polynomial b){
-    if(exp==1){
+Polynomial Polynomial::pow_mod(Polynomial base, long long exp, IrreduciblePolynomial p, Polynomial b){
+    //cout << "pm(" << base.value() << ", " << exp << ", " << p.value() << ", " << b.value() << '\n';
+    if(exp==1LL){
         return base*b % p;
-    }else if(exp%2==0){
+    }else if(exp%2==0LL){
         return pow_mod(base*base % p, exp/2, p, b);
     }
     return pow_mod(base*base % p, (exp-1)/2, p, base*b % p);
